requirements.txt
python-telegram-bot==13.15
schedule
requests
python-binance